Big, big refactor
- Renamed `MonsterType` to `CreatureType`
- `ObjectType` values now know the `ArtemisObject` subclasses they
produce
- All the various world object update packets have been collapsed into a
single `ObjectUpdatePacket`.
- Methods annotated with `@Listener` can now listen for updates to
`ArtemisObjects` or their subtypes.
(`ArtemisAnomaly`) and support the new upgrade mechanism.
- Fixed problem with `ByteArrayReader` where reading a `BitField` did
not move the pointer.
- Added new `PlayerShipUpdateListener` convenience class for listening
for updates to an `ArtemisPlayer` with a specific ship index.
- New `ArtemisAnomaly` and `ArtemisCreature` classes; `ArtemisWhale`
removed.
- `BaseArtemisObject` and its subclasses now have simplified
constructors that only expect an object ID (as that is the only property
that is guaranteed to be available
- `ArtemisGenericObject` will throw an exception if you attempt to use
it for an inappropriate ObjectType.